fix: some request properties are not being passed to hawk in Node16
Due to https://github.com/nodejs/node/issues/36550 (it was reverted in
Node 14 but it is back in Node 16). req.headers and other properties are
not own properties of requests so they are not cloned.

Node documentation (see discussion on https://github.com/nodejs/node/issues/36550)
advices against clonning the req (or any stream object). This PR fixes that
without modifying the original request object by creating a new object which
prototype is the original request object instead of trying to clone
it.